v3.4.4
Added dependency ‘Blinker’ to handle signals such as ‘template_rendered’ which trigger execution of a callback function. This, in conjunction with Flask, adds the ability to record trial_start timestamps when a template is rendered. To install, use pip install blinker==1.9.0
Dependency was added to startup.bat.
Added ‘observed_urls’ and ‘record_path’ to settings. This a list of URL paths where timestamps will be recorded, and whether to include the path in the timestamp entry sent to the database, respectively (see below).
Resolved issue where timestamps were not being recorded correctly
Timestamps recorded did not correspond to actual trial timing - Moved timestamp capture outside of the ‘/trial/{n}’ route function. Seems like timestamps were being captured at GET and POST HTTP requests and thus overwrote the actual values. Recorded timestamps will also include the url of the page there were captured in. Timestamps are now captured in the following manner:
Trial_start is captured in the callback function for the ‘template_rendered’ signal (Blinker.Signal) if the URL of the rendered page matches (starts with) any of the urls defined in ‘observed_urls’.
Trial_end is captured in the callback function decorated with ‘@app.before_request’ using the same condition.
Captured timestamps were timezone-localised to the local timezone instead of the timezone defined in settings. This was resolved by using the zoneinfo library alongside datetime.
Added a second method to retrieve the next_story_index parameter from the database. Rather than counting the number of individual distinct entries in that column, the database can now take the maximum value from the next_story_index column (typecasted as an integer).
Fixed an issue where the app would ignore the selected number of stories to view in a session, or would show that number of stories +1. This was done by comparing the max_story_index parameter against next_story_index rather than current_story_index (idk why it was set this way tbh).
Fixed an issue where the app would crash when trying to create a new data directory for a new generated user id because the directory was already created somewhere in the route instead of within ‘write_userdata_to_file()’.
Added template ‘final_end.html’ alongside a matching route to redirect users if they have completed all stories in their story order. This fixes the app crashing when trying to access an index in story_order which is out of range.
Created the command line argument ‘-v’ or ‘--verbose’ so that the administrator can decide whether they want verbose detailed console outputs regardless of what mode the app is run on.
Minor changes:
Remove pytz module, replaced with zoneinfo
Changed comment, ‘data’ to ‘session’ line 808.